+++ steamGamesStarter@KopfdesDaemons/files/steamGamesStarter@KopfdesDaemons/desklet.js
@@ -9,8 +9,14 @@ const DESKLET_DIR = imports.ui.deskletManager.deskletMeta[UUID].path;
 
 imports.searchPath.push(DESKLET_DIR);
 
-const SteamHelper = imports.helpers.steam.SteamHelper;
-const UiHelper = imports.helpers.ui.UiHelper;
+let SteamHelper, UiHelper;
+if (typeof require !== "undefined") {
+  SteamHelper = require("./helpers/steam.js").SteamHelper;
+  UiHelper = require("./helpers/ui.js").UiHelper;
+} else {
+  SteamHelper = imports.helpers.steam.SteamHelper;
+  UiHelper = imports.helpers.ui.UiHelper;
+}
 
 Gettext.bindtextdomain(UUID, GLib.get_home_dir() + "/.local/share/locale");